feat(191): allow proto3 and newer editions
This should allow for yet unreleased editions (> 2024) without needing
any bumps to the linter.

Despite proto3 being part of the editions enum, the IsProto3() still
needs to be checked to reliably detect proto3 syntax.

The practical result of this rule should be that anything newer than
proto2 is now allowed.